+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `ComputeProximityScores` methods to compute global proximity scores for cell graphs.
 - `heuristic_illumination` to approximate natural illumination of a 3D layout.
+- Illumination masking option in `render_rotating_layout` to modulate `node_val` colors with a heuristic illumination mask.
 - "NaturalBlue" option as a gradient palette in `PixelgenGradient` to use as a natural illuminated color gradient.
 
 ### Updated